anses_agent_crewai/pyproject.toml
[project]
name = "anses-agent"
version = "1.0.0"
description = "Agente ANSES para emision de constancias CUIL. Sistema A2A."
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.5",
    "crewai[tools]>=0.126.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "google-generativeai>=0.8.0",
]
